requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
joblib
pymongo
python-dotenv
fastapi
uvicorn